requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.9
h11==0.14.0
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
pillow==11.1.0
psycopg2==2.9.10
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyJWT==2.10.1
pyotp==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.20
requests==2.32.3
setuptools==75.8.2
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wheel==0.45.1
